fix some errors

+++ geopositionmap/geoFormFields.py
@@ -33,8 +33,8 @@ class LatLngFormField(forms.MultiValueField):
     def __init__(self, *args, **kwargs):
         #self.widget = geoWidgets.LatLngTextInputWidget()
         fields = (
-            forms.DecimalField(label=_('latitude')),
-            forms.DecimalField(label=_('longitude')),
+            forms.FloatField(label=_('latitude')),
+            forms.FloatField(label=_('longitude')),
         )
         if 'initial' in kwargs:
             kwargs['initial'] = LatLng(*kwargs['initial']).pos
